security.SELinux: Ensure the specified file exists.
CheckContext's behavior was confusing: existence of the specified
file was not checked when recursive=false, while it was checked
when recursive=true. This is also not optimal because the test
did not fail when a file is moved (crbug.com/933582 is such an
example).

This CL updates CheckContext to always check existence of files.

BUG=chromium:1009309
TEST=fast_build.sh -T
TEST=tast run betty security.SELinuxFilesSystem security.SELinuxFilesARC

Change-Id: If63312a1421859c9330ee6436b086f468f4b0d66
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1832191
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Qijiang Fan <fqj@google.com>