tast-tests: Proxy adb command status code.
Even if the execution in ARC container fails, "adb exec-out" does not
return it to the caller.
This CL replaces it by "adb shell". "adb exec-out" is introduced to
avoid automatic convert between LF and CRLF, however nowadays
it is properly handled in adb.

This CL also adds a sanity test to ensure the behavior.

BUG=chromium:1008804
TEST=Ran locally.

Change-Id: Iefe344e07f71eb56e8a89eb753f9a3bed36f5e19