tast/video: Use media-internals for hw accelerator usage check
This CL adds a new function checkIfURLUsesPlatformVideoDecoder()
that digs into chrome://media-internals to find out if a given
URL is being played back with is_platform_video_decoder set or
not -- which is equivalent to hw decoding acceleration or not.

With that, this CL removes the logic for testing histogram entry
reading before-after, and only checks histograms for errors.

Callsites are updated to not specify if they wanted to check the
histograms or not: this is because during the transition (see
go/vd-migration, tests with "vd" in their names) the "new"
implementation and the "old" use different histograms.  This CL
changes the M.O. to looking at chrome:media-internals always,
which should be ok for both cases; moreover, this CL changes to
checking for no errors in all cases.

checkIfURLUsesPlatformVideoDecoder() is based on JS that gets
injected into chrome://media-internals; this JS is in a new file
that is added to the appropriate Data dependencies.

BUG=chromium:1009335
TEST=tast -verbose run -build=true 100.127.97.100 video.Play* on nocturne

Change-Id: I8d859e38b7c8948b7c3f5993ff5ec65d48cb2a58