tast/video: Use media-internals for hw accelerator usage check
This CL adds a new function checkIfURLUsesPlatformVideoDecoder()
that digs into chrome://media-internals to find out if a given
URL is being played back with is_platform_video_decoder set or
not -- which is equivalent to hw decoding acceleration or not.

Callsites are updated to pass play.(No)VerifyHWAcceleratorUsed
to TestPlay(); when set to verify, TestPlay() will load the JS
code in the new file chrome_media_internals_utils.js and use it
to figure out if the video was played using an accelerator; the
error histograms are also checked in this case just like ToT
does. This CL removes the logic for reading use-histograms
before and after the test logic since it's not needed anymore.

BUG=chromium:1009335
TEST=tast -verbose run -build=true 100.127.97.100 video.Play* on nocturne

Change-Id: I8d859e38b7c8948b7c3f5993ff5ec65d48cb2a58