Update patch set 3
Patch Set 3:

(1 comment)

Thank you for preparing the change!

A high level remark, as I had no time to diff the changes:

The main reason for the hasty version of glbench to exist is to protect the non-hasty version which is too slow to run in the CQ. If there is no non-hasty version in tast (only running in autotest) then this tast test doesn't have a real purpose in life. In other words landing this change depends on what we do with long running graphics tests. Now running long tast tests (especially such that can reboot/hang machines) depends on how well tast supports machine recovery (currently questionable). So there is no hurry to land this at the moment. As in the email discussion this needs to be followed up by a doc, which I am preparing (but I am traveling).

Patch-set: 3
CC: Gerrit User 1337146 <1337146@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>