Tast: assistant tests clean-up.
This change cleans up the temporary logic for parsing Assistant
response, after the API change CL has been submitted.
Also because the time query, previously used for checking Libassistant
service readiness, is not working as expected for now, this change
replaces it with a simple math query, which serves the same purpose
and should be more reliable.

BUG=None.
TEST=./fast_build.sh -T
     tast -verbose run {DUT_ip} ui.AssistantSimpleQueries

Change-Id: Ia92f9719fdd1a9a26888f24d76656d10ed3eb3df
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1836536
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Tested-by: Meilin Wang <meilinw@chromium.org>
Commit-Queue: Meilin Wang <meilinw@chromium.org>