Add testPIPExpandViaMenuTouch and testPIPExpandViaShelfIcon
This CL adds the sub-tests for two of the scenarios where PIP gets
expanded.
The remaining work is to add tests that expand PIP with alt-tab,
but it probably requires a new autotest private API as we need to
robustly distinguish short alt-tab and long alt-tab and injecting
key events to achieve this can be flaky.

BUG=b:142029367
TEST=tast -verbose run my_crbook arc.PIP

Change-Id: I9edff3efae37a88f810ddd865a6b789bde2db9d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1836302
Commit-Queue: Kazuki Takise <takise@chromium.org>
Tested-by: Kazuki Takise <takise@chromium.org>
Auto-Submit: Kazuki Takise <takise@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>