Tast: retrieve WiFi interface from shill.
Add GetWifiInterface() in shill/wifi.go to get a managed WiFi interface
from shill. Previously, the WiFi interface was hard-coded as "wlan0" and
the test fails on certain devices. Now, test can dynamicall get the
system's WiFi interface to test to solve the issue.

BUG=chromium:988894,chromium:999707
TEST=tast run $DUT network.IWScan on elm, which was failed because its
wifi interface is "mlan0", not the hard-coded "wlan0".

Change-Id: Iacbd7ebe35affd3dd61e518f378a0adc27bb25be
Signed-off-by: Shuo-Peng Liao <deanliao@chromium.org>