fluffy_interop_std: Use NewProxy for Servod
Since servod may be running on a machine that is not localhost, we would
need to open up a SSH tunnel to talk to servod.  This commit simply
replaces the method of creating the servo object with NewProxy (instead
of just New) which will open up a SSH connection to the servod host if
needed.

BUG=None
TEST=Run fluffy_interop_std against a fluffy connected to a labstation
from my workstation; verify it can connect okay.

Change-Id: I98353162c4cc7bff3c6d267d83a4c5502033dad2
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1837457
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>
Auto-Submit: Aseda Aboagye <aaboagye@chromium.org>