fluffy_interop_std: Use NewProxy for Servod
Since servod may be running on a machine that is not localhost, we would
need to open up a SSH tunnel to talk to servod.  This commit simply
replaces the method of creating the servo object with NewProxy (instead
of just New) which will open up a SSH connection to the servod host if
needed.

BUG=None
TEST=Run fluffy_interop_std against a fluffy connected to a labstation
from my workstation; verify it can connect okay.

Change-Id: I98353162c4cc7bff3c6d267d83a4c5502033dad2
Signed-off-by: Aseda Aboagye <aaboagye@google.com>