Tast-tests: Add arc.MultiDisplay(Visibility with maximized window) #2
Add #2 subtest for arc.MultiDisplay. The test is created based on
b/110105532. When an activity is maximized on one display, another
activity on other display is not visible and shows as black. This test
is to make sure it doesn't happen and the activity still shows at the
same place.

Bug: 134981068
Test: connect one extra display and run "tast -verbose run DUT
      arc.MultiDisplay"

Change-Id: I2a428012398033d1aa11ab367ade42d51daf59ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1823959
Tested-by: Cici Ruan <cuicuiruan@google.com>
Auto-Submit: Cici Ruan <cuicuiruan@google.com>
Commit-Queue: Cici Ruan <cuicuiruan@google.com>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>