webrtc.MediaRecorderPerf: Increase timeout of media recorder perf tests.
This CL increases the timeout of the webrtc media recorder performance tests. On
successful runs some devices seem to take 2m25+ for the test to complete. This
is too close to the configured timeout of 3 minutes, as some time is reserved
for test cleanup.

The timeout is increased from 3 minutes to 5 minutes to leave some margin to
prevent future flakiness.

TEST=webrtc.MediaRecorderPerf on samus

BUG=b:140359777

Change-Id: I395b775f13cdd6d4f97292e084c3e6a42e40ef3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1839313
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Tested-by: David Staessens <dstaessens@chromium.org>
Commit-Queue: David Staessens <dstaessens@chromium.org>