Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> I think Dean's cl (https://crrev.com/c/1819140) will fix the same issue, so this cl can be abandoned.
> However, this change reminds me that Billy added a 'IwlwifiPCIRescan' test in an abandoned parent commit (https://crrev.com/c/1727573).
> Want to make sure if that test is still necessary?

We may drop its GetWiFiInterface() fix. However, it has a fix in iw_runner.go::ListInterfaces(), which is a different story. The function is not in ToT yet so it must be in its dependent pending CL. You can find it and amend the fix. After that we can abandon the CL.

Patch-set: 4