Update patch set 2
Patch Set 2:

(3 comments)

> Patch Set 2:
> 
> For DefaultProfile and DefaultProfileServices, I think the only failures I see are related to real bugs or other similar issues. e.g., https://crbug.com/1006026. So those are probably fine to promote.
> 
> Also, to note why Tastboard didn't notify about the TwoShills failures: it looks like it only looks at 'master' branch failures -- it doesn't notice release-* branch failures. (That's a bit worrisome.) It also doesn't highlight stuff older than a few weeks, and one of the master-branch failures was right about 1 month ago.

If the master branch tests pass while release-* branches' tests failure, that might imply something is fixed in master not cherry-picking back to older release branch yet, right? And CL promoting some test to critical only affects master branch, it won't promote release branch' test unless if we cherry-pick the CL to release branch. If the above is correct, it should be okay to promote a test when we see the test is green on master for a while, say 20 builds.

For TwoShills test, I agree with your comment. It is not ready yet. Thanks for pointing it out.

Patch-set: 2
Reviewer: Gerrit User 1001853 <1001853@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>