Tast: promote two network Tasts to critical.
network.DefaultProfile and  DefaultProfileServices were green for 20
builds.
Note that though IWGetSet is green for 20 build, as the test itself is
quite imcomplete, I'd rather keep it in informational.
For TwoShills, it is still flaky on release branch and has some issues
waiting to be fixed. Let's keep it in information for now.

BUG=chromium:1011597
TEST=None

Change-Id: Ibc9fea730b44bb5c5448c9c381e7bcf216c531ae