Remove WaitForIdle in WindowManagerCUJ test.
This CL replaces usages of WaitForIdle in the WindowManagerCUJ test.
This should fix flakiness, particularly around the PIP test. This CL
also adds some TODOs for areas of the code that test for 'nothing
happening' which we currently do not have a good solution for doing
robustly.

BUG=chromium:972897
TEST=tast run dut-1 arc.WindowManagerCUJ

Change-Id: I45088c5ddca903f032d9a59fbc82286dde614811
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1846959
Tested-by: Eliot Courtney <edcourtney@chromium.org>
Commit-Queue: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>