tast/video: Use media-internals for perf tests hw accelerator usage
This CL follows up on crrev.com/c/1832413 and teaches playback.go
to use chrome://media-internals to check if the video decode
accelerator was used in a given test or not, eliminating the
need to rely on histogram updates.

The current logic in play.go is extracted into a new file called
data/chrome_media_internals.js; this forces an update in the
play_decode_accel_used*.go files.

BUG=chromium:1009335, chromium:1009890
TEST=tast -verbose run -build=true papaya video.PlayDecodeA* video.Playback*
on nocturne

Change-Id: Id8aa088a8997856ea93dbcdfb179c97614f6aca7