video: disable VP9.2 tests using hw acceleration
This CL disables the tests that use VP9.2 hardware acceleration.
This is because the current Chrome implementation does not support
HDR not colorspace-specific conversions, and for the time being
we'll be just using sw decoding.

The sister CL crrev.com/c/1842564 disables enumerating this codec,
forcing the use of sw decoding.

BUG=chromium:1011469, chromium:1011454
TEST=tast -build=true 100.127.97.100 video.*92* on nocturne

Change-Id: Ia7b4af0945ae1d77bdc663934375ac8150ba871f