check_disk_space: Increase required disk space margin
As we are seeing build failures due to the rootfs growing larger than the
filesystem on a more regular basis, it would be helpful if there was an
earlier warning. If we increase the margin from 2%/11MB to 5%/40MB we are
more likely to catch unexpected rootfs growth early and be able to take
action before build failures start.

BUG=chromium:998345
TEST=None

Change-Id: I8fc5fde082912e994852a3dd50333106f6deae3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1850012
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>