tast-tests: arc.WindowManagerCUJ use new .apk version
The new .apk does not relies on Companion Library, making it portable to
arc-next with any additional change.
That means that the new .apk does not report the device and window
state.
This is not a problem since that information can be fetched from Tast.

This CL just makes the needed changes to use the new .apk.

TEST=tast run $DUT arc.WindowManagerCUJ
BUG=b:142499827

Change-Id: I79b9418ce8da5dcb9eada1ebf82ee41b33b93b43