Update patch set 5
Patch Set 5:

(1 comment)

So splitting the test into stress and non-stress variants seems like a good thing to do and looks good, but it doesn't seem to have anything to do with the patch subject -- do we need to do that alongside the extra checks for some reason or can we split the patch so those can go in separately (and not be hidden to people doing e.g. git log --oneline)?

Patch-set: 5
Reviewer: Gerrit User 1117055 <1117055@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>