Rename WaitForIdle to WaitForResumed.
WaitForIdle seems a bit too general and has been used incorrectly. This
CL renames it to what it actually does, which is waiting for an activity
to be resumed (in P, at least)WaitForIdle seems a bit too general and
has been used incorrectly. This CL renames it to what it actually does,
which is waiting for an activity to be resumed (in P, at least).

BUG=None
TEST=tast run dut-1 arc.*

Change-Id: Id469fbdadfb1b607c4c0e97583f1f4ac6bb846f0