tast/video: Increase running time for video.PlayDecodeAccelUsed.*
This CL increases slightly the running time left for the videos
to play in pollPlaybackCurrentTime, from 0.9 s to 1.5s: this
is because chrome:media-internals is updated every second or so.
Crosbolt runs hasn't seen this flake but it showed up in my
local hammer runs, so let's increase this minimally.

BUG=chromium:1009335, chromium:1009890
TEST=for i in {1..10}; do tast -verbose run papaya video.PlayDecodeAccelUsed.*; done
on kohaku

Change-Id: I1df1db375d5c8a384e4798f9810acbff3f277f3f