tast/webrtc: Use getStats for {En,De}codeAccelUsedVP8
This CL migrates the tests webrtc.{Decode,encode}AccelUsedVP8
to use RTCPeerConnection getStats() API instead of checking
histograms.

It takes the chance to:
- Remove the use of a y4m file, which is not needed to exercise
  an RTCPeerConnection and slow the test down.
- Move the supporting logic to video/peer_connection.go,
  separating it from the support code for DecodeAccelUsedJPEG
  (this is completely different and is left untouched waiting
  for further cleanup, it should be under the camera/ folder).

Note that getStats() doesn't fill in the codec implementation
immediately, so we Poll() for up to one second. Experimentally,
it takes ~2 or 3 VideoFrames to decide that value.

BUG=chromium:1012828
TEST=tast -verbose run papaya webrtc.*AccelUsed*

Change-Id: I127697bd899e5d112f748cf4661d319291204c82