tast-tests: arc.DisplayRotationPerf more reliable
This CL introduces "display.WaitForDisplayRotation()" function, which
waits until the rotation animation is finished. This makes the
tests that relies on screen rotation less flaky.

This CL also updates arc.DisplayRotationPerf by adding a call to this
new function.

And finally, this CL fixes some nits regarding the indentation in the
embedded JS code in display.go

BUG=b:73009846
TEST=tast run $DUT arc.ScreenRotationPerf
     Visally noticed that the wait for the screen rotation is more
     stable.

Change-Id: I7ee146c96a88a00ac8d41c7f3d827c9c67af2ce7