arc.ResizeActivity: Wait for window stable before move window
The test failure is hard to reproduce locally.

When moving window, it needs to use window bounds to get the swipe start
coordinates and the end coordinates, and check if the coordinates are
inside of the display bounds. The test failed due to the swipe coordinate
position is out of bounds. The window bounds are not stable after
resizing the window. Improve the test to wait until the window stable
and then move the window.

BUG=b:150731172
TEST=tast -verbose run DUT arc.ResizeActivity

Change-Id: I97a5d91a0c5dd6606bcee2939155164d1c2f4199