Chrome crash tests: Avoid PID-reuse race
Avoid failing Chrome crash tests because of a PID-reuse race. In particular,
ignore any processes created after we get the initial list of processes; if
a subsequent process gets the PID we were using when we kill it, we'll still
count the process as killed.

BUG=chromium:1012980
TEST=tast -verbose run --extrauseflags chrome_internal my_crbook ui.ChromeCrashLoggedIn

Change-Id: Ia734181489279403dd26dedc8066b3df90452065
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1865694
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Ian Barkley-Yeung <iby@chromium.org>