Wait for activity to be visible on the Chrome side in arc.WindowState.
Previously, it only waited until the activity was resumed. I observed on
some runs on Tastboard that the test was getting to the part where it
checks the window state before the window was available on the Chrome
side. This change makes the test wait until the activity is known to
Chrome before proceeding to the main body of the test.

BUG=None
TEST=tast run dut-1 arc.WindowState.*

Change-Id: Ie3d0b5bf4fe6919baf1fe3e6dd119e956243316c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1863185
Reviewed-by: Prameet Shah <phshah@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Prameet Shah <phshah@chromium.org>
Tested-by: Prameet Shah <phshah@chromium.org>