wilco.ECEvent: Introduce test
ECEvent tests the Wilco EC is able to generate EC events and that the
kernel properly passes them along to userspace to read. Normally, the EC
triggers events on hardware changes, such as battery errors. For
testing, we can send a special command to the EC, and the EC will
generate a hardcoded  dummy event.

BUG=b:134059465
TEST=tast run $DUT wilco.ECEvent

Change-Id: Id282584a3d153840f1daf607af8ef6c68061da90