Update patch set 26
Patch Set 26: Code-Review-1

Hi, sorry for overlooking your comment directed to me. I usually don't look at +2'ed CLs, but I should change this routine.

I have to say that testing.Sleep is not an acceptable option, especially when we don't have any concrete plan to fix it. It's not because it's "The Rule" for Tast tests, but it's because it's a well known bad practice for all functional tests in general. According to a study which Keigo shared earlier in tast-users@, sleep is the most popular source of test flakiness:
https://groups.google.com/a/chromium.org/d/topic/tast-users/KUofbGPw_gI/discussion

I have two possible ideas to attack this problem.

Firstly, if the problem is about waiting for the "first" udev event after restart, how about watching "udevadm monitor" output? The steps are:

1. Run "udevadm settle" to make sure there's no active event prior to tests.
2. Start "udevadm monitor" process.
3. Restart udev or restart wifi device.
4. Read at least one line of stdout of the udevadm process launched in step 2.
5. Kill the udevadm process.
6. Run "udevadm settle" again.

There will be still races obviously, but the point here is to increase the probability of test failure as much as possible when the product is actually broken, while avoiding false positives (that is, test failure when the product is functional).

Secondly, can we split the assertion into two steps? The former step would be to assert that the list of interfaces is changed from the original one, and the latter step would be to assert the the list is reverted back to the original one. The current approach is difficult because we are trying to assert that a state (list of interfaces) does not change. I guess splitting assertion is possible for the WiFi interface restart case, but I'm not sure if it's possible for the udev restart case.

I hope this helps.

Patch-set: 26
Reviewer: Gerrit User 1118346 <1118346@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1