tast/video: Simplify video.Seek.* tests
This CL takes a stab at trying to fix the video.Seek.* tests,
that sometimes get borked and prevent other subsequent Tast
tests from running.  These tests play a video and do a lot of
seeks in it; a seek is started via setting <video>.currentTime,
a seeking event is fired upon start and a seeked event is
fired at the end of the seek.

This CL does a couple of things:

- The video resolution (specially the seek_switch_* cases) can
  be very large, which I think might be too much for some DUTs,
  and is beside the point of this test; this CL forces the
  <video> to be <=50% of the tab width,

- The <video> element desn't have the 'loop' attribute, so it
  could perfectly finish, failing subsequently to seek.

- Part of the ToT logic consists on seeking to without waiting
  for the seek to finish (they are called "fast-seek"); the HTML
  media elements ignore any seeks while one is in progress, so
  this part is unnecessary and this CL removes it; instead this
  CL utilizes many seek-seeked cycles.

Finally video.html had unused code that is removed.

BUG=chromium:1012030
TEST=tast -verbose=true papaya video.Seek.* on kohaku

Change-Id: Ieab37b8c1ff80ea026abf47c36c3af1829cac223