Break out Chrome crash tests
In
https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1846411,
I added tests that killed various processes within the Chrome process
tree. In an effort to save time, I brought Chrome up once and then
killed various processes within that one bring up.

Well, it looks like that was a mistake. Tastboard shows that frequently,
no broker process exists by the time the tests try to kill the broker.
(Example:
https://stainless.corp.google.com/browse/chromeos-autotest-results/swarming-47f13427e7b18810/autoserv_test/tast/results/tests/ui.ChromeCrashLoggedIn/)
I guess it's not surprising that after Chrome is killed and has to restart
its pieces, it's not always in a good state. It's not ideal, but it's
really not what this test is supposed to be testing.

Be less fancy and just recreate Chrome before each test by logging out
and starting Chrome up again.

BUG=chromium:984807
TEST=tast -verbose run --extrauseflags chrome_internal my_crbook ui.ChromeCrashNotLoggedIn* ui.ChromeCrashLoggedIn*

Change-Id: I7f4d49ea160b8034ce07d377ff356caf95ac6db3