Update patch set 2
Patch Set 2:

nya@ I was looking into b/142828365 and ADB a little more and this is an alternative to 
crrev.com/c/1866149 that might actually be a bit simpler overall.

Note that b/37104408 was an ADB bug for the race when doing "adb kill-server; adb start-server" but as we have seen even with this "fix" the race still exists. (I'm not exactly sure why the fix did not work.)

After thinking about it, the connection lifetimes are technically independent of the process lifetime anyway (eg TIME_WAIT state). With killing and waiting for a non-child process being tricky I think it might be simpler to handle the occasionally "connection reset" error and retry the connection. If we were actually connecting to a remote system and not loopback then this how we would solve it instead of trying reason about the process states.

Patch-set: 2
Reviewer: Gerrit User 1118346 <1118346@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1115919 <1115919@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1308136 <1308136@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>