tast/video: Simplify Seek.*: limit max <video> size and make it loop
This CL takes a stab at trying to fix the video.Seek.* tests,
that sometimes get borked and prevent other subsequent Tast
tests from running.  These tests play a video and do a lot of
seeks in it; a seek is started via setting <video>.currentTime,
a seeking event is fired upon start and a seeked event is
fired at the end of the seek.

This CL modifies the <video> element in the HTML:

- The video resolution (specially the seek_switch_* cases) can
  be very large, which I think might be too much for some DUTs,
  and is beside the point of this test; this CL forces the
  <video> to be <=50% of the tab width,

- The <video> element desn't have the 'loop' attribute, so it
  could perfectly finish, failing subsequently to seek.

BUG=chromium:1012030
TEST=tast -verbose=true run papaya video.Seek.* on kohaku

Change-Id: Ic8526567590075d691e6bc6e4dde962a406faa0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1869334
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Tested-by: Miguel Casas <mcasas@chromium.org>
Commit-Queue: Ilja H. Friedel <ihf@chromium.org>