RELAND: tast/video: Simplify Seek.*: remove fast-seeks
The original CL crrev.com/c/1869337 was reverted because it
broke the bots somehow (test didn't finish on time). This CL
reduces the amount of seeks, which were increased in the
reverted CL.

Original CL description --------------------------------------
This CL takes a stab at trying to fix the video.Seek.* tests,
that sometimes get borked and prevent other subsequent Tast
tests from running.  These tests play a video and do a lot of
seeks in it; a seek is started via setting <video>.currentTime,
a seeking event is fired upon start and a seeked event is
fired at the end of the seek.

Part of the ToT logic consists on seeking to without waiting
for the seek to finish (they are called "fast-seek"); the HTML
media elements ignore any seeks while one is in progress, so
this part is unnecessary and this CL removes it; instead this
CL utilizes many seek-seeked cycles.

BUG=chromium:1012030
TEST=tast -verbose=true run papaya video.Seek.* on kohaku

Change-Id: I54b5acc4956bd4a713cf54ae17d7182392078153