tast/video: Simplify video.Seek.* tests, part 3/3
This CL takes a stab at trying to fix the video.Seek.* tests,
that sometimes get borked and prevent other subsequent Tast
tests from running.  These tests play a video and do a lot of
seeks in it; a seek is started via setting <video>.currentTime,
a seeking event is fired upon start and a seeked event is
fired at the end of the seek.

Part of the ToT logic consists on seeking to without waiting
for the seek to finish (they are called "fast-seek"); the HTML
media elements ignore any seeks while one is in progress, so
this part is unnecessary and this CL removes it; instead this
CL utilizes many seek-seeked cycles.

BUG=chromium:1012030
TEST=tast -verbose=true run papaya video.Seek.* on kohaku

Change-Id: I00c1b193a166ba290437c227ca58f52c61deca63