tast/video: Simplify Seek.*: remove fast-seeks
This CL takes a stab at trying to fix the video.Seek.* tests,
that sometimes get borked and prevent other subsequent Tast
tests from running.  These tests play a video and do a lot of
seeks in it; a seek is started via setting <video>.currentTime,
a seeking event is fired upon start and a seeked event is
fired at the end of the seek.

Part of the ToT logic consists on seeking to without waiting
for the seek to finish (they are called "fast-seek"); the HTML
media elements ignore any seeks while one is in progress, so
this part is unnecessary and this CL removes it; instead this
CL utilizes many seek-seeked cycles.

BUG=chromium:1012030
TEST=tast -verbose=true run papaya video.Seek.* on kohaku

Change-Id: I00c1b193a166ba290437c227ca58f52c61deca63
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1869337
Tested-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Miguel Casas <mcasas@chromium.org>