Update patch set 3
Patch Set 3:

(2 comments)

I unfortunately know very, very, very little about ACPI objects. It would probably be wise to get someone who actually knows something about them to review this.

My mostly-clueless thoughts: this is a very fuzzy replacement for proper parsing of the ASL. It looks like you might match various sorts of arbitrary nesting patterns, which may or may not be the correct object. Is there really no better way to get this sort of object information programmatically? e.g., it doesn't show up in sysfs in any better form than raw AML?

I'd also think we might want to explicitly check which WiFi driver is in use. As it is, this test might be misleading, pretending it can check anything that's not Intel WiFi. It would be better to have a clear message for that case.

And, if ath10k goes smoothly, we'll have a much easier time parsing / testing it -- https://crbug.com/968289 -- so it might be nice to leave the door open to testing other vendors.

By the way, it'll be bad enough having this test around, but not run regularly. So we probably won't notice it breaking in arbitrary ways. It would be best to make its failure modes (e.g., "wrong vendor") as clear as possible.

Patch-set: 3
Reviewer: Gerrit User 1132414 <1132414@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>