Add perf related logic in cca.go and collect them in CCAUIPreviewPerf
To measure the performance for basic usage of CCA, we should add some
perf related logic in order to set the connection between CCA background
and Tast test. In addition, we collect those perf events in
CCAUIPreviewPerf.

The corresponding CL on Chrome side is:
https://crrev.com/c/1880995
We should land that CL (Chrome) first before landing this CL (Tast).

BUG=b:141518806
TEST=tast run [DUT] camera.CCAUIPreviewPerf

Change-Id: Ic47bc7915d560ec83b1699103715473b2b1cfb57