tast-test: fix WaitUntilIdle strategy in ui.OverviewPerf
Individual invocation of WaitUntilIdle may wait for 1 minute at
most, and it would take too much time for waiting until idle for
individual operation. Currently it waits for 4 times, it might
wait upto 4 minutes. Since timeout is set to 2 minutes in this
test, it will fail as timeout in such case.

Rather than increasing the timeout, I believe we can also switch
the timing of waiting, IMO we don't have to wait for switching
between clamshell and the tablet mode. This way it invokes
WaitUntilIdle 2 times. To allow doing the remaining tasks, the
timeout is increased to 2m30s.

Also, if CPU might not be idle after 1 minute passes, I don't
think it should stop processing. It would be fine to proceed,
but it should record the waiting. Thus modified to use s.Log().

BUG=chromium:1016943
TEST=tast run ui.OverviewPerf

Change-Id: I9b54c5493f5dde653731f93e15bb37424361f945