video.PlaybackPerf: Remove video name suffix from reported perf metrics.
Currently the name of the video file used to test is reported to crosbolt as
part of the metric names. But this information is actually duplicated, as the
video name is already incorperated into the test name. This means that when
looking up test results in crosbolt we will see entries such as:
tast.video.PlaybackPerfH2641080P30FPS: tast_hw_video_cpu_usage_h264_1080p_30fps

This CL changes this to:
tast.video.PlaybackPerfH2641080P30FPS: tast_hw_video_cpu_usage

This also means the performance metric names are now in sync with those used by
the video.DecodeAccelPerf tests.

These are the new metric names:
- tast_hw_video_cpu_usage
- tast_hw_video_dropped_frames
- tast_hw_video_dropped_frames_percent
- tast_hw_video_power_consumption
- tast_sw_video_cpu_usage
- tast_sw_video_dropped_frames
- tast_sw_video_dropped_frames_percent
- tast_sw_video_power_consumption

TEST=video.PlaybackPerfH2641080P30FPS on nocturne

BUG=None

Change-Id: Ifeb8dcbaf9e3e4913fcefd2da477d862e94cf91c