tast-tests: crostini: Add support for testing the buster image
We add support for specifying that we want to test the buster image for
the tast tests. This is a temporary measure, eventually buster will
become the default image and we can swap all the tast tests to use it.

This CL also contains a sample tast test, crostini.SanityDownloadBuster,
which demonstrates the usage. In a future CL we will add buster versions
for all relevant tests.

BUG=chromium:930901
TEST=tast run $DUT crostini.*Buster

Change-Id: I8c3968c4708fe899565dc29bbd29af9f435125a2