chrome: GetRootPID as session_manager direct child
The process walking performed in GetPIDs() and GetRootPID() suffer from
inherent races in the /proc interface. The process states can change both
while the filtering and the ppid relationship inspection is happening. The
end result is that the logic in GetRootPID(), while correct for a tree,
may be operating on a tree with missing nodes and in which there is no
single root.

Sidestep the problem by retrieving the session_manager PID and identifying
the chrome process that is a direct child of session_manager as the root
browser process.

BUG=b:142974350
TEST=tast run IP arc.Boot.stress

Change-Id: Ifc2bd32f0079288065eb087a08e8f26d26d69e6c