chrome: Avoid hung in browserWatcher.close.
browserWatcher was tricky to use because browserWatcher.close freezes
if (1) it is called twice or (2) it is called before start is called.

Actually (1) was true when browserWatcher.close returns an error in
logInAsGuest. This change fixes the issue.

Also this change avoids possible similar issues caused by (2) by
calling start in newBrowserWatcher, which removes the unstarted state
entirely.

BUG=chromium:1018014
BUG=chromium:1018034
TEST=fast_build.sh -T
TEST=tast run atlas security.UserFilesGuest  # 100+ times

Change-Id: I4cb82122e2d2a15afb5137f1304c53a784fdefd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1880551
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>