chrome: Wait for the browser process before restarting.
We used to ignore GetRootPID's error, but it can return an error
if session_manager hasn't started the browser process yet. In that
case, oldPID is set to -1, so the following testing.Poll can break
as soon as the non-testing browser process comes up. This confuses
browserWatcher because it might recognize the non-testing browser
process as the testing browser process and report a false browser
crash.

BUG=chromium:1018014
BUG=chromium:1018034
TEST=fast_build.sh -T
TEST=tast run atlas security.UserFilesGuest  # 100+ times

Change-Id: Id3a5ceb2e4cabc41b77a2903c924256d5d884aa5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1880553
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>