tast/video: decrease amount of seeks for Seek.* tests
This CL extends the Seek.* parameters to define the total amount
of seeks to be performed and reduces this number from 100 to 25,
since last runs have shown that some small chromebooks have
troubles performing so many and time out [1] (and it's unlikely
that a device does e.g. 30 seeks and then chokes).

[1] https://stainless.corp.google.com/search?view=list&first_date=2019-10-19&last_date=2019-10-25&test=%5Etast.video.Seek.*&build=R80-12621.0.0&status=WARN&status=FAIL&status=ERROR&status=NOT_RUN&exclude_cts=false&exclude_not_run=false&exclude_non_release=true&exclude_au=true&exclude_acts=true&exclude_retried=false&exclude_non_production=false

BUG=chromium:1018045
TEST=tast -verbose run papaya video.Seek.* on kohaku

Change-Id: I399fc2b69a9566a77e18d147ba5b94bdd7f4aea3