tast: Force ALSA tests to test on internal mic and speaker
CL:1862362 changed the logic of device name reported from CRAS. It
causes the queried name may not be equal to the device which really
used CRAS. It makes audio.ALSAConformance and audio.Microphone select
the wrong device to test.

To fix it, we can force them to test on internal mic and speaker, which
we make sure the names are still correct now. After we get the label
supported and correct PCM name, we can create the different tast to test
external audio devices.

BUG=b:142910355
TEST=tast run {IP} audio.ALSAConformance audio.Microphone

Change-Id: Idcd53aff356dccf7c2fadd71d31862708062ee31
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1880211
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Commit-Queue: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>